requirements-docker.txt
bumpversion
celery
click
coreapi
django-admin-autocomplete-filter
django-admin-numeric-filter
django-bestzhlocale
django-cors-headers
django-debug-toolbar
django-dj-plugin>=0.3.0
django-environ
django-extensions
django-filter
django-kronos
django-middleware-global-request
django-openauth
django-pandas
django>2,<3
djangorestframework-jwt
djangorestframework>=3.8.2
gunicorn[gevent]
markdown
psycopg2-binary
pygments
pytelegraf
pytest
pytest-django
pyyaml
sentry-sdk
whitenoise
xlwt
